UDT-23 refactor: Project Enum Converter Package(service -> util)
