+++ src/main/java/com/example/udtbe/global/util/CategoryTypeConverter.java
@@ -1,4 +1,4 @@
-package com.example.udtbe.global.service;
+package com.example.udtbe.global.util;
 
 import com.example.udtbe.entity.enums.CategoryType;
 import jakarta.persistence.AttributeConverter;
